refactor structure
