+++ backend/src/auth/utils.py
@@ -1,0 +1,3 @@
+from .config import pwd_context
+def verify(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
