+++ backend/src/config.py
@@ -1,0 +1,8 @@
+class Auth:
+    max_username_size=50
+    max_password_size=200
+class App:
+    DSN="https://4001ffe917ccb261aa0e0c34026dc343@o4505702629834752.ingest.us.sentry.io/4507694792704000"
+    TRACES_SAMPLE_RATE=1.0
+    PROFILES_SAMPLE_RATE=1.0
+    LOCALHOST="http://localhost:8080"
