+++ backend/src/database.py
@@ -1,0 +1,8 @@
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy.ext.declarative import declarative_base
+
+Base = declarative_base()
+engine = create_engine("sqlite:///news_database.db", echo=True)
+Base.metadata.create_all(engine)
+Session = sessionmaker(bind=engine)
