+++ backend/src/news/config.py
@@ -1,0 +1,1 @@
+GPT_MODEL="gpt-3.5-turbo"
