+++ backend/src/news/router.py
@@ -1,0 +1,131 @@
+import itertools
+import requests
+import json
+from fastapi import Depends, FastAPI
+from openai import OpenAI
+from bs4 import BeautifulSoup
+from ..auth.dependencies import session_opener,authenticate_user_token
+from ..models import NewsArticle
+from .service import get_article_upvote_details, get_new_info, toggle_upvote
+from .schemas import PromptRequest, NewsSumaryRequestSchema
+from .config import GPT_MODEL
+app = FastAPI()
+@app.get("/api/v1/news/news")
+def read_news(db=Depends(session_opener)):
+    """
+    read new
+
+    :param db:
+    :return:
+    """
+    news = db.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
+    result = []
+    for article in news:
+        upvotes, upvoted = get_article_upvote_details(article.id, None, db)
+        result.append(
+            {**article.__dict__, "upvotes": upvotes, "is_upvoted": upvoted}
+        )
+    return result
+
+@app.get("/api/v1/news/user_news")
+def read_user_news(
+        db=Depends(session_opener),
+        usertoken=Depends(authenticate_user_token)
+):
+    """
+    read user new
+    :param db:
+    :param u:
+    :return:
+    """
+    news = db.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
+    result = []
+    for article in news:
+        upvotes, upvoted = get_article_upvote_details(article.id, usertoken.id, db)
+        result.append(
+            {
+                **article.__dict__,
+                "upvotes": upvotes,
+                "is_upvoted": upvoted,
+            }
+        )
+    return result
+
+_id_counter = itertools.count(start=1000000)
+@app.post("/api/v1/news/search_news")
+async def search_news(request: PromptRequest):
+    prompt = request.prompt
+    news_list = []
+    ai_info = [
+        {
+            "role": "system",
+            "content": "你是一個關鍵字提取機器人，用戶將會輸入一段文字，表示其希望看見的新聞內容，請提取出用戶希望看見的關鍵字，請截取最重要的關鍵字即可，避免出現「新聞」、「資訊」等混淆搜尋引擎的字詞。(僅須回答關鍵字，若有多個關鍵字，請以空格分隔)",
+        },
+        {"role": "user", "content": f"{prompt}"},
+    ]
+    completion = OpenAI(api_key="xxx").chat.completions.create(
+        model=GPT_MODEL,
+        messages=ai_info,
+    )
+    keywords = completion.choices[0].message.content
+    # should change into simple factory pattern
+    news_items = get_new_info(keywords, is_initial=False)
+    for news in news_items:
+        try:
+            response = requests.get(news["titleLink"])
+            item_soup = BeautifulSoup(response.text, "html.parser")
+            # 標題
+            item_title = item_soup.find("h1", class_="article-content__title").text
+            item_time = item_soup.find("time", class_="article-content__time").text
+            # 定位到包含文章内容的 <section>
+            content_section = item_soup.find("section", class_="article-content__editor")
+
+            paragraphs = [
+                paragraphinfo.text
+                for paragraphinfo in content_section.find_all("p")
+                if paragraphinfo.text.strip() != "" and "▪" not in paragraphinfo.text
+            ]
+            detailed_news = {
+                "url": news["titleLink"],
+                "title": item_title,
+                "time": item_time,
+                "content": paragraphs,
+            }
+            detailed_news["content"] = " ".join(detailed_news["content"])
+            detailed_news["id"] = next(_id_counter)
+            news_list.append(detailed_news)
+        except Exception as error:
+            print(error)
+    return sorted(news_list, key=lambda x: x["time"], reverse=True)
+
+@app.post("/api/v1/news/news_summary")
+async def news_summary(
+        payload: NewsSumaryRequestSchema, user=Depends(authenticate_user_token)
+):
+    response = {}
+    ai_info = [
+        {
+            "role": "system",
+            "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
+        },
+        {"role": "user", "content": f"{payload.content}"},
+    ]
+    completion = OpenAI(api_key="xxx").chat.completions.create(
+        model=GPT_MODEL,
+        messages=ai_info,
+    )
+    result = completion.choices[0].message.content
+    if result:
+        result = json.loads(result)
+        response["summary"] = result["影響"]
+        response["reason"] = result["原因"]
+    return response
+
+@app.post("/api/v1/news/{id}/upvote")
+def upvote_article(
+        id,
+        db=Depends(session_opener),
+        usertoken=Depends(authenticate_user_token),
+):
+    message = toggle_upvote(id, usertoken.id, db)
+    return {"message": message}
