+++ backend/src/prices/router.py
@@ -1,0 +1,10 @@
+from fastapi import Query
+import requests
+@app.get("/api/v1/prices/necessities-price")
+def get_necessities_prices(
+        category=Query(None), commodity=Query(None)
+):
+    return requests.get(
+        "https://opendata.ey.gov.tw/api/ConsumerProtection/NecessitiesPrice",
+        params={"CategoryName": category, "Name": commodity},
+    ).json()
